Refine UI gradients and styling for visual consistency
- Update KPI cards to use pink-dominant gradient matching brand
- Apply light lavender background to AI Goal Coach panel
- Reduce navigation link font size for cleaner header
- Switch numeric parsing to parseFloat for decimal support
- Unify progress bar and circle gradients across pages

+++ backend/src/services/parser.js
@@ -82,7 +82,7 @@ export async function parseAnnualObjectives(filePath) {
       const targetMatch = line.match(/Target\*\*: (.+)/);
       if (targetMatch) {
         const val = targetMatch[1].trim();
-        currentKR.target = isNaN(val) ? 0 : parseInt(val);
+        currentKR.target = isNaN(val) ? 0 : parseFloat(val);
       }
     }
 
@@ -90,7 +90,7 @@ export async function parseAnnualObjectives(filePath) {
     else if (line.startsWith('- **Current**:') && currentKR) {
       const currentMatch = line.match(/Current\*\*: (.+)/);
       if (currentMatch) {
-        currentKR.current = parseInt(currentMatch[1].trim()) || 0;
+        currentKR.current = parseFloat(currentMatch[1].trim()) || 0;
       }
     }
